fix: quantity의 데이터타입 int -> Long 변경

+++ src/main/java/gift/wishlist/dto/WishRequestDto.java
@@ -9,7 +9,7 @@ public record WishRequestDto(
         Long productId,
 
         @Min(value = 1, message = "수량은 1개 이상이어야 합니다.")
-        int quantity
+        Long quantity
 ) {
         public Wish toEntity(Long memberId) {
                 return new Wish(memberId, this.productId, this.quantity);
